Simplify absence status system to two states only
- Remove 'approved' status, keeping only 'reported' and 'pending verification'
- 'reported' absences are now effectively accepted (green CheckCircle icon)
- 'pending verification' for cases requiring manual review (amber Warning icon)
- Update backend API validation to enforce new two-status constraint
- Add database constraint limiting status values
- Update frontend components to reflect simplified status workflow

+++ backend/database/database.js
@@ -109,8 +109,8 @@ class Database {
             const query = `
                 SELECT 
                     class_name,
-                    COUNT(*) as absence_count,
-                    COUNT(DISTINCT student_name) as unique_students,
+                    COUNT(*)::int as absence_count,
+                    COUNT(DISTINCT student_name)::int as unique_students,
                     STRING_AGG(DISTINCT reason, ', ') as reasons
                 FROM absences 
                 WHERE DATE(absence_date) BETWEEN DATE($1) AND DATE($2)
